\name{MVR.news}
\alias{MVR.news}
\title{Function to Display the NEWS File}
\description{
    Function to display the NEWS file of the \pkg{MVR} package.
}
\usage{
    MVR.news(...)
}
\arguments{
    \item{...}{Further arguments passed to or from other methods.}
}
\value{
    None.
}
\author{
    \itemize{
        \item Jean-Eudes Dazard, Ph.D. \email{jxd101@case.edu}
        \item Hua Xu, Ph.D. \email{hxx58@case.edu}
        \item Alberto Santana, MBA. \email{ahs4@case.edu}
        \item J. Sunil Rao, Ph.D. \email{JRao@med.miami.edu}
    }
    Maintainer: Jean-Eudes Dazard, Ph.D. \email{jxd101@case.edu}
    
    Acknowledgments: This project was partially funded by the 
    National Institutes of Health (P30-CA043703 to J-E.DAZARD, R01-GM085205 to J.S.RAO), 
    and the National Science Foundation (DMS-0806076 to J.S.RAO).
}
\note{
    End-user function. 
}
\keyword{documentation}
